+++ pkg/storegateway/indexheader/lazy_binary_reader_test.go
@@ -14,12 +14,15 @@ import (
 	"time"
 
 	"github.com/go-kit/log"
+	"github.com/grafana/dskit/gate"
 	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
 	promtestutil "github.com/prometheus/client_golang/prometheus/testutil"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
 	"github.com/thanos-io/objstore"
+	"go.uber.org/atomic"
 
 	"github.com/thanos-io/objstore/providers/filesystem"
 
@@ -294,6 +297,76 @@ func testLazyBinaryReader(t *testing.T, bkt objstore.BucketReader, dir string, i
 		return NewStreamBinaryReader(ctx, logger, bkt, dir, id, 3, NewStreamBinaryReaderMetrics(nil), Config{})
 	}
 
-	reader, err := NewLazyBinaryReader(ctx, factory, logger, bkt, dir, id, NewLazyBinaryReaderMetrics(nil), nil)
+	reader, err := NewLazyBinaryReader(ctx, factory, logger, bkt, dir, id, NewLazyBinaryReaderMetrics(nil), nil, gate.NewNoop())
 	test(t, reader, err)
+}
+
+// TestLazyBinaryReader_ShouldBlockMaxConcurrency tests if LazyBinaryReader blocks
+// concurrent loads such that it doesn't pass the configured maximum.
+func TestLazyBinaryReader_ShouldBlockMaxConcurrency(t *testing.T) {
+	ctx := context.Background()
+
+	tmpDir := filepath.Join(t.TempDir(), "test-indexheader")
+	bkt, err := filesystem.NewBucket(filepath.Join(tmpDir, "bkt"))
+	require.NoError(t, err)
+	t.Cleanup(func() { require.NoError(t, bkt.Close()) })
+
+	// Create block.
+	blockID, err := block.CreateBlock(ctx, tmpDir, []labels.Labels{
+		labels.FromStrings("a", "1"),
+		labels.FromStrings("a", "2"),
+		labels.FromStrings("a", "3"),
+	}, 100, 0, 1000, labels.FromStrings("ext1", "1"))
+	require.NoError(t, err)
+	require.NoError(t, block.Upload(ctx, log.NewNopLogger(), bkt, filepath.Join(tmpDir, blockID.String()), nil))
+
+	logger := log.NewNopLogger()
+
+	const (
+		numLazyReader          = 20
+		maxLazyLoadConcurrency = 3
+	)
+
+	var (
+		totalLoaded = atomic.NewUint32(0)
+		inflight    = atomic.NewUint32(0)
+	)
+
+	errOhNo := errors.New("oh no")
+
+	factory := func() (Reader, error) {
+		testInflight := inflight.Inc()
+		require.LessOrEqual(t, testInflight, uint32(maxLazyLoadConcurrency))
+		totalLoaded.Inc()
+
+		time.Sleep(3 * time.Second)
+
+		inflight.Dec()
+
+		return nil, errOhNo
+	}
+
+	var lazyReaders [numLazyReader]*LazyBinaryReader
+	lazyLoadingGate := gate.NewInstrumented(prometheus.NewRegistry(), maxLazyLoadConcurrency, gate.NewBlocking(maxLazyLoadConcurrency))
+
+	for i := 0; i < numLazyReader; i++ {
+		lazyReaders[i], err = NewLazyBinaryReader(ctx, factory, logger, bkt, tmpDir, blockID, NewLazyBinaryReaderMetrics(nil), nil, lazyLoadingGate)
+		require.NoError(t, err)
+	}
+
+	var wg sync.WaitGroup
+	wg.Add(numLazyReader)
+
+	// Attempt to concurrently load 20 index-headers.
+	for i := 0; i < numLazyReader; i++ {
+		index := i
+		go func() {
+			_, err := lazyReaders[index].IndexVersion()
+			require.ErrorIs(t, err, errOhNo)
+			wg.Done()
+		}()
+	}
+
+	wg.Wait()
+	require.Equal(t, totalLoaded.Load(), uint32(numLazyReader))
 }